requirements.txt
Django==4.0.4
django-phonenumber-field==6.1.0
phonenumbers==8.12.47
Pillow==9.1.0
environs==9.5.0
requests==2.27.1
aiogram==2.20
djangorestframework==3.13.1
markdown==3.4.1
django-filter==22.1